Фаза света: один запрос динамики вместо девяноста шести, + разметка кадра
Разбор фазы света довёл до корня: 90% её времени -- не отрисовка, а ожидание.
Прибор da_light_prof разложил её на части (Юпитер, безлимит теневых ламп):

  ожидание списков  1.25 мс      теневые карты 0.07      накопление 0.07

Ожидание оказалось нехваткой контекстов: 96 обходов сцены за кадр на 15
рисуемых ламп, 27 остановок главного потока, из них 26 -- по нехватке.
Отключение сбора статики (r__light_skip_static 2) не изменило НИЧЕГО, а
отключение запроса динамики уронило ожидание до 0.14 -- вся стоимость в
SpatialSpace.q_frustum, по одному запросу к дереву объектов на лампу.

Теперь запрос ОДИН на всю фазу: объём по сферам влияния всех теневых ламп,
дальше каждая лампа фильтрует готовый список своей пирамидой тем же тестом,
что применялся внутри дерева. Замер в одном прогоне: 1.30 -> 0.39 мс, состав
теней не изменился (15 ламп, 16 накоплений, 96 обходов). Выключатель
r__light_dyn_shared 0.

Ни один публичный форк этого не делает: CoC, Monolith и IX-Ray спрашивают
пирамидой КАМЕРЫ (и теряют тени от объектов вне поля зрения), OpenXRay, OGSR
и Refined -- пирамидой лампы, но из цикла никто запрос не выносил.

Разметка кадра доведена до конца: prepare / occq / wmarks / tail плюс
вложенные sky, forward, sorted, wait_fence, wait_cull. Раньше зоны покрывали
только проходы отрисовки, и половина процессорного кадра в отчёт не попадала.
На Кордоне так нашлось 1.3-2.7 мс простоя в r_main.sync() -- ожидание расчёта
видимости сцены, при нуле работы видеокарты и нуле вызовов отрисовки.

Отсечение частиц по расстоянию (r__particle_dist, 150 м): 0.72 -> 0.26 мс,
эффектов в кадре 585 -> 186. Авторское отсечение из ветки OpenGL не годилось:
порог там привязан к дальности видимости мира, а точка отсчёта верна лишь для
части эффектов.

Порог мелких объектов выведен наружу (r__ssa_discard) и поднят 3.5 -> 4.5:
элементов 3090 -> 2966, проходов 533 -> 505. Отраслевая практика отсекает
всё, что уже 4-8 пикселей; 3.5 стояло ниже этого диапазона и не менялось.

Вылет на критическом попадании в сталкера закрыт с двух концов: пустой список
анимаций и непроверенные слот оружия, тип ранения и активный предмет. Четыре
VERIFY, которых в релизе нет, заменены живыми отказами; play_global пустой
MotionID обрабатывает штатно.

Разбор очереди отрисовки: da_geom_dump и da_forward_dump. Склейка вызовов по
диапазонам индексов оценена и отклонена -- 580 вызовов из 2966 при 0.58 мкс
на вызов, ниже нижней отраслевой границы.

+++ src/xrGame/script_game_object.h
@@ -577,7 +577,8 @@ public:
     void set_patrol_path(LPCSTR path_name, const EPatrolStartType patrol_start_type,
         const EPatrolRouteType patrol_route_type, bool random);
     void inactualize_patrol_path();
-    void set_dest_level_vertex_id(u32 level_vertex_id);
+    // [DA_PORT] Именно u64: скрипт присылает сюда 4294967296 — см. .cpp
+    void set_dest_level_vertex_id(u64 level_vertex_id);
     void set_dest_game_vertex_id(GameGraph::_GRAPH_ID game_vertex_id);
     void set_movement_selection_type(ESelectionType selection_type);
     u32 level_vertex_id() const;
